requirements.txt
psycopg2-binary>=2.9.9
openai>=1.0.0
Pillow>=10.0.0
mitmproxy>=10.0.0
psutil>=5.9.0
python-dotenv>=1.0.0
requests>=2.31.0
frida>=16.0.0
frida-tools>=12.0.0
